# frontend-React/openapi-v1.yaml
openapi: 3.0.3
info:
  title: GECE API
  version: 1.0.0
servers:
  - url: https://api.example.com/api/v1
paths:
...
  /projects:
    get:
      summary: List projects
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
    post:
      summary: Create project
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /projects/{project_id}:
    get:
      summary: Get project by id
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: project_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        tenant_id: { type: string }
        status: { type: string }
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }